// client/src/modules/product/types/product.ts
export interface ProductImage {
  id: string;
  url: string;
  alt: string;
  label?: string;
}

export interface ProductColor {
  id: string;
  colorName: string;
  colorHex: string;
  displayImage: string;
  mainImage: string;
  galleryImages: string[];
  sizes?: string[];
}

export interface ProductColorVariation {
  id: string;
  colorName: string;
  colorHex: string;
  size?: string;
  sizeName?: string;
  thumbnail: string;
  price: number;
  originalPrice: number;
  discountPercentage: number;
  sku: string;
  stock: number;
  images: ProductImage[];
}